app/backend/requirements.txt
asgiref==3.4.1
Django==3.2.16
djangorestframework==3.14.0
pkg-resources==0.0.0
psycopg2==2.9.5
python-dotenv==0.20.0
pytz==2022.6
sqlparse==0.4.3
typing-extensions==4.1.1
